+++ src/app/store.js
@@ -1,6 +1,6 @@
 import {configureStore} from "@reduxjs/toolkit";
 import authReducer from "../features/auth/authSlice";
-import usersReducer from "../features/ProfilePage/userSlice";
+import usersReducer from "../features/ProfilePages/userSlice";
 
 export const store = configureStore({
     reducer: {
