+++ src/features/index/index.js
@@ -1,5 +1,6 @@
 export { Feed } from "../feed/Feed";
 export {Explore} from "../explore/Explore"; 
-export {ProfilePage} from "../ProfilePage/ProfilePage";
+export {ProfilePage} from "../ProfilePages/ProfilePage";
+export {OtherUsersProfile} from "../ProfilePages/OtherUsersProfile";
 export {BookMarks} from "../bookmarks/BookMarks";
 export {Login, SignUp, } from "../auth";
